refactor: 출력관련 테스트 ConsoleOutputHandlerTest 클래스로 이동
