+++ src/test/java/lotto/domain/result/LottoStatisticsTest.java
@@ -4,7 +4,6 @@ import lotto.domain.number.TestLottoNumberFactory;
 import lotto.domain.ticket.LottoTicket;
 import lotto.domain.ticket.LottoTickets;
 import lotto.domain.ticket.WinningLotto;
-import lotto.io.ConsoleOutputHandler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,12 +11,11 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 import java.util.Map;
 
-import static lotto.io.ConsoleOutputHandler.PROFITABLE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoStatisticsTest {
 
-    public static final int LOTTO_TICKET_PRICE = 1000;
+    private static final int LOTTO_TICKET_PRICE = 1000;
     private LottoTickets lottoTickets;
     private WinningLotto winningLotto;
 
@@ -57,12 +55,4 @@ class LottoStatisticsTest {
         assertThat(statistics.getProfitRatio()).isGreaterThan(1.0);
     }
 
-    @Test
-    @DisplayName("수익률에 따른 평가 메시지를 확인할 수 있다.")
-    void getAssessmentBasedOnProfitRatio() {
-        LottoStatistics statistics = lottoTickets.generateLottoStatistics(winningLotto, LOTTO_TICKET_PRICE);
-        ConsoleOutputHandler consoleOutputHandler = new ConsoleOutputHandler();
-
-        assertThat(consoleOutputHandler.getAssessmentText(statistics.getProfitRatio())).contains(PROFITABLE);
-    }
 }
